fix: Restructured project folders and deleted unwanted files
