+++ imports/ui/styles/tailwind.css
